Implemented layers. Fixed rendering of many sprites (> 16k). GL errors are now printed out as strings.

+++ Bengine-core/src/graphics/buffers/indexbuffer.cpp
@@ -4,12 +4,12 @@ namespace bengine
 {
 	namespace graphics
 	{
-		IndexBuffer::IndexBuffer(GLushort * data, GLsizei count)
+		IndexBuffer::IndexBuffer(GLuint * data, GLsizei count)
 			: _count(count)
 		{
 			glGenBuffers(1, &_bufferId);
 			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, _bufferId);
-			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), data, GL_STATIC_DRAW);
+			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLuint), data, GL_STATIC_DRAW);
 			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 		}
