+++ Bengine-core/src/graphics/buffers/indexbuffer.h
@@ -13,7 +13,7 @@ namespace bengine
 			GLuint _count;
 
 		public:
-			IndexBuffer(GLushort * data, GLsizei count);
+			IndexBuffer(GLuint * data, GLsizei count);
 			~IndexBuffer();
 
 			void bind() const;
